requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.5.2
gunicorn==21.2.0
pytest==7.4.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
SQLAlchemy==2.0.25
alembic==1.13.1
pandas==2.1.3